refactor: standardize PascalCase in sections, adjust import, update structure and documentation
