+++ Struture.md
@@ -1,95 +1,88 @@
-# Visão Geral (meta) — Estrutura Feature-based
+# Estrutura Feature-based — Guia para Devs
 
-Objetivo: agrupar tudo relacionado ao "feature resume" (dados, resolução de assets, hooks e componentes) em uma estrutura coesa, testável e fácil de navegar; manter `components/ui` para primitives reutilizáveis; reduzir acoplamento entre provider, transformação e UI.
+Este projeto utiliza uma estrutura **feature-based** para facilitar manutenção, testes e evolução do código.
 
-## Por que isso importa
+## Objetivo
 
-- Facilita encontrar e modificar o código relacionado ao currículo (dados + UI).
-- Separa responsabilidades: resolução de assets, transformação de dados, provisionamento (Provider) e apresentação (components).
-- Torna mais simples migrar a origem dos dados (por exemplo trocar `resume.json` por uma API) sem tocar na UI.
+- Agrupar tudo relacionado ao currículo (dados, assets, hooks, componentes) em uma estrutura coesa e fácil de navegar.
+- Manter `components/ui` para primitives reutilizáveis.
+- Reduzir acoplamento entre provider, transformação e UI.
 
-## Design Pattern aplicado
+## Vantagens
 
-Organização de código / estrutura de pastas
+- Encontrar e modificar o código do currículo fica mais rápido.
+- Separação clara de responsabilidades: assets, dados, provider, UI.
+- Facilita migrar a origem dos dados (ex: trocar `resume.json` por API) sem mexer na UI.
 
-- Feature-based: agrupar por recurso/feature (ex.: `components/sections/experience/*`, `components/ui/*`). O projeto já está organizado dessa forma; a sugestão é manter e reforçar essa convenção.
-- Cada pasta de feature deve conter, quando aplicável:
+## Padrão de Organização
 
-  - `index.ts` (barrel export) — exporta a API pública da pasta;
-  - `ComponentName.tsx` — componente principal (PascalCase);
-  - `ComponentName.module.css` — estilos em CSS Modules;
-  - `types.ts` — tipos específicos da feature;
+- **Feature-based:** agrupe por recurso/feature (ex: `components/sections/experience/*`, `components/ui/*`).
+- Cada pasta de feature pode conter:
+  - `index.ts` (barrel export) — exporta a API pública da pasta
+  - `ComponentName.tsx` — componente principal (PascalCase)
+  - `ComponentName.module.css` — estilos em CSS Modules
+  - `types.ts` — tipos específicos da feature
 
-- Convenção de nomes: componentes em PascalCase (`AboutSection.tsx`), hooks com prefixo `use` (`useResume.ts`), arquivos de serviço em camelCase (`assetResolver.ts`).
+### Convenção de nomes
 
-Por que usar esse padrão
+- Componentes: PascalCase (`AboutSection.tsx`)
+- Hooks: prefixo `use` (`useResume.ts`)
+- Serviços: camelCase (`assetResolver.ts`)
 
-- Reduz imports relativos longos e melhora navegação.
-- Facilita mover ou reescrever uma feature inteira sem tocar outras partes do app.
-- Barrels (`index.ts`) tornam os imports mais limpos e explícitos.
+## Por que usar esse padrão?
 
-## Estrutura proposta (feature-based)
+- Imports mais curtos e navegação facilitada.
+- Possível mover/refatorar uma feature sem afetar o resto do app.
+- Barrels (`index.ts`) deixam os imports limpos e explícitos.
+
+## Exemplo de Estrutura
 
 ```txt
 src/
   features/
     resume/
       context/
-        DataProvider.tsx       # Provider refatorado a partir de DataContext.tsx
-        index.ts               # exporta Provider e hooks públicos
+        DataProvider.tsx       # Provider do contexto
+        index.ts               # Barrel export
       services/
-        assetResolver.ts       # lógica de import.meta.glob e resolveAssetReference
-        transformResume.ts     # função de traverse/transform do objeto resume
+        assetResolver.ts       # Resolve assets via import.meta.glob
+        transformResume.ts     # Transforma dados do resume
       hooks/
-        useResume.ts           # hook simples que consome o context
-        useResumeSelector.ts   # (opcional) selector para reduzir re-renders
+        useResume.ts           # Hook principal
+        useResumeSelector.ts   # (opcional) selector para otimizar re-render
       types/
-        resume.ts              # tipos exportados (Resume)
+        resume.ts              # Tipos do currículo
       components/
-        ResumeContainer.tsx    # container opcional que injeta props nas sections
+        ResumeContainer.tsx    # Container opcional para injetar props
   components/
-    ui/                      # primitives reutilizáveis (buttons, avatar, etc.)
+    ui/                      # Primitives reutilizáveis (buttons, avatar, etc.)
   data/
     resume.json
   assets/
 ```
 
-## Passo a passo resumido
+## Passo a Passo para Refatoração
 
-- [x] 1. Criar `src/features/resume/services/assetResolver.ts`
+1. Criar `src/features/resume/services/assetResolver.ts` (mover lógica de assets)
+2. Criar `src/features/resume/services/transformResume.ts` (função para transformar dados)
+3. Refatorar `DataContext.tsx` para `DataProvider.tsx` e exportar via barrel
+4. Mover hook `useResume` para pasta de hooks
+5. Adicionar barrels (`index.ts`) nas subpastas
+6. Atualizar imports em `main.tsx` e `App.tsx`
+7. (Opcional) Criar `ResumeContainer.tsx` e tornar sections apresentacionais
+8. Rodar `npm run build` e validar assets/favicon
 
-  - mover a lógica de `import.meta.glob` e exportar `resolveAsset`.
+## Riscos e Assunções
 
-- [x] 2. Criar `src/features/resume/services/transformResume.ts`
+- Não há SSR (Server Side Rendering) — se houver, proteja efeitos que usam `document`/`window`.
+- `import.meta.glob` é específico do Vite.
 
-  - implementar a função que clona `resume.json` e substitui referências de imagem por URLs usando `resolveAsset`.
+## Dicas Finais
 
-- [x] 3. Criar `src/features/resume/context/DataProvider.tsx` e `context/index.ts`
+- Reorganizar melhora navegabilidade e desacoplamento.
+- Mudanças são de baixo risco e reversíveis.
+- Atualizações de imports podem ser automatizadas por busca/replace.
 
-  - mover e simplificar `DataContext.tsx` para `DataProvider` que usa `transformResume` e fornece o contexto.
-  - exportar `DataProvider` e `useResume` no `context/index.ts`.
+---
 
-- [x] 4.  Criar `src/features/resume/hooks/useResume.ts`
-
-- mover o hook atual `useResume` para aqui e exportar via `context/index.ts`.
-
-- [x] 5.  Adicionar barrels (`index.ts`) nas subpastas e no feature root
-
-- atualizar imports em `src/main.tsx` e `src/App.tsx` para apontar para `features/resume` quando aplicável.
-
-- [x] 6.  (Opcional) Criar `ResumeContainer.tsx` e tornar as sections apresentacionais
-
-- refatorar seções para receber props ao invés de chamar `useResume` diretamente.
-
-- [x] 7.  Rodar build/checar aplicação
-- executar `npm run build` e abrir a versão local para garantir que assets e favicon foram resolvidos corretamente.
-
-## Riscos e assunções
-
-- Assumi que não há SSR — se houver, efeitos que usam `document`/`window` precisam de guards adicionais.
-- `import.meta.glob` é Vite-specific; o `assetResolver` manterá essa dependência.
-
-## Observações finais
-
-- A reorganização melhora navegabilidade e desacoplamento. É um conjunto de mudanças de baixo risco e facilmente reversível; atualizações dos imports podem ser automatizadas por uma simples busca/replace.
-- Se quiser que eu aplique essa migração automaticamente, posso começar movendo `DataContext.tsx` e criando `assetResolver` e `transformResume` agora, atualizando os imports e validando a build.
+Se quiser aplicar essa migração automaticamente, posso iniciar movendo arquivos e atualizando imports para você!
