+++ src/components/sections/ProjectsSection.tsx
